build memkind optionally
