+++ benchmark/CMakeLists.txt
@@ -7,7 +7,11 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR}/../libjiffy/src
         ${Boost_INCLUDE_DIRS})
 
 add_executable(hash_table_bench src/hash_table_benchmark.cpp)
-add_dependencies(hash_table_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(hash_table_bench boost_ep memkind_ep)
+else()
+	add_dependencies(hash_table_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(hash_table_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -16,7 +20,11 @@ install(TARGETS hash_table_bench
 
 add_executable(file_bench src/file_benchmark.cpp)
 
-add_dependencies(file_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(file_bench boost_ep memkind_ep)
+else()
+	add_dependencies(file_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(file_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -34,7 +42,11 @@ install(TARGETS shared_log_bench
 
 add_executable(fifo_queue_bench src/fifo_queue_benchmark.cpp)
 
-add_dependencies(fifo_queue_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(fifo_queue_bench boost_ep memkind_ep)
+else()
+	add_dependencies(fifo_queue_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(fifo_queue_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -43,7 +55,11 @@ install(TARGETS fifo_queue_bench
 
 add_executable(file_auto_scaling_bench src/file_auto_scaling_benchmark.cpp)
 
-add_dependencies(file_auto_scaling_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(file_auto_scaling_bench boost_ep memkind_ep)
+else()
+	add_dependencies(file_auto_scaling_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(file_auto_scaling_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -52,7 +68,11 @@ install(TARGETS file_auto_scaling_bench
 
 add_executable(fifo_queue_auto_scaling_bench src/fifo_queue_auto_scaling_benchmark.cpp)
 
-add_dependencies(fifo_queue_auto_scaling_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(fifo_queue_auto_scaling_bench boost_ep memkind_ep)
+else()
+	add_dependencies(fifo_queue_auto_scaling_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(fifo_queue_auto_scaling_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -61,7 +81,11 @@ install(TARGETS fifo_queue_auto_scaling_bench
 
 add_executable(hash_table_auto_scaling_bench src/hash_table_auto_scaling_benchmark.cpp src/zipf_generator.h)
 
-add_dependencies(hash_table_auto_scaling_bench boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(hash_table_auto_scaling_bench boost_ep memkind_ep)
+else()
+	add_dependencies(hash_table_auto_scaling_bench boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(hash_table_auto_scaling_bench jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
@@ -70,7 +94,11 @@ install(TARGETS hash_table_auto_scaling_bench
 
 add_executable(hash_table_auto_scaling_get src/hash_table_auto_scaling_get.cpp src/zipf_generator.h)
 
-add_dependencies(hash_table_auto_scaling_get boost_ep memkind_ep)
+if (BUILD_MEMKIND_SUPPORT)
+	add_dependencies(hash_table_auto_scaling_get boost_ep memkind_ep)
+else()
+	add_dependencies(hash_table_auto_scaling_get boost_ep jemalloc_ep)
+endif()
 
 target_link_libraries(hash_table_auto_scaling_get jiffy_client ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
