+++ libjiffy/bench/file_block_benchmark.cpp
@@ -6,9 +6,9 @@
 #include <jiffy/utils/logger.h>
 #include <jiffy/utils/signal_handling.h>
 #include <jiffy/utils/time_utils.h>
+#include <jiffy/utils/mem_utils.h>
 #include "jiffy/storage/file/file_defs.h"
 #include "jiffy/storage/file/file_partition.h"
-#include <memkind.h>
 
 using namespace ::jiffy::client;
 using namespace ::jiffy::directory;
@@ -46,20 +46,13 @@ int main(int argc, char const *argv[])
     }
     
     else if (vm.count("dram")) {
-        struct memkind* pmem_kind = nullptr;
         std::string memory_mode = "DRAM";
     }
-    if (vm.count("pmem")) {
-        struct memkind* pmem_kind = nullptr;
+    else if (vm.count("pmem")) {
         std::string memory_mode = "PMEM";
         std::string pmem_path = vm["pmem"].as<std::string>();
-        size_t err = memkind_create_pmem(pmem_path.c_str(),0,&pmem_kind);
-        if (err) {
-            char error_message[MEMKIND_ERROR_MESSAGE_SIZE];
-            memkind_error_message(err, error_message, MEMKIND_ERROR_MESSAGE_SIZE);
-            fprintf(stderr, "%s\n", error_message);
-        }
     }
+    void* pmem_kind = mem_utils::init_kind(memory_mode, pmem_path);
 
     std::string address = "127.0.0.1";
     int service_port = 9090;
@@ -122,5 +115,4 @@ int main(int argc, char const *argv[])
 	LOG(log_level::info) << "\t" << data_size << " payload";
 	LOG(log_level::info) << "\tThroughput: " << num_ops * 1E3 / tot_time << " requests per microsecond";
     
-    memkind_destroy_kind(pmem_kind);
 }
