+++ libjiffy/src/jiffy/storage/block.h
@@ -4,7 +4,6 @@
 #include <string>
 #include <vector>
 #include <jiffy/utils/property_map.h>
-#include <memkind.h>
 #include "chain_module.h"
 #include "partition.h"
 
@@ -23,7 +22,7 @@ class block {
   explicit block(const std::string &id,
         const size_t capacity = 134217728,
         const std::string memory_mode = "DRAM",
-        struct memkind* pmem_kind = nullptr,
+        void* pmem_kind = nullptr,
         const std::string &auto_scaling_host = "127.0.0.1",
         const int auto_scaling_port = 9095);
 
@@ -82,7 +81,7 @@ class block {
  private:
   std::string id_;
   block_memory_manager manager_;
-  struct memkind* pmem_kind_;
+  void* pmem_kind_;
   std::shared_ptr<chain_module> impl_;
 
   std::string directory_host_;
