+++ libjiffy/src/jiffy/storage/block_memory_manager.cpp
@@ -1,4 +1,8 @@
-#include <memkind.h>
+#ifdef MEMKIND_IN_USE
+  #include <memkind.h>
+#else
+  #include <jemalloc/jemalloc.h>
+#endif
 #include <new>
 #include "block_memory_manager.h"
 #include "jiffy/utils/logger.h"
@@ -7,29 +11,42 @@ using namespace jiffy::utils;
 namespace jiffy {
 namespace storage {
 
-block_memory_manager::block_memory_manager(size_t capacity, const std::string memory_mode, struct memkind *pmem_kind)
+block_memory_manager::block_memory_manager(size_t capacity, const std::string memory_mode, void* pmem_kind)
     : capacity_(capacity), used_(0), memory_mode_(memory_mode), pmem_kind_(pmem_kind) {}
 
 void *block_memory_manager::mb_malloc(size_t size) {
   if (used_.load() > capacity_) {
     return nullptr;
   }
-  if (memory_mode_ == "DRAM") {
-    pmem_kind_ = MEMKIND_DEFAULT;
-  }
-  auto ptr = memkind_malloc(pmem_kind_, size);
+  #ifdef MEMKIND_IN_USE
+    if (memory_mode_ == "DRAM") {
+      pmem_kind_ = MEMKIND_DEFAULT;
+    }
+    auto ptr = memkind_malloc((struct memkind*)pmem_kind_, size);
+  #else
+    auto ptr = mallocx(size, 0);
+  #endif 
   used_ += size;
   return ptr;
 }
 
 void block_memory_manager::mb_free(void *ptr) {
-  auto size = memkind_malloc_usable_size(pmem_kind_, ptr);
-  memkind_free(pmem_kind_, ptr);
+  #ifdef MEMKIND_IN_USE
+    auto size = memkind_malloc_usable_size((struct memkind*)pmem_kind_, ptr);
+    memkind_free((struct memkind*)pmem_kind_, ptr);
+  #else 
+    auto size = sallocx(ptr, 0);
+    free(ptr);
+  #endif
   used_ -= size;
 }
 
 void block_memory_manager::mb_free(void *ptr, size_t size) {
-  memkind_free(pmem_kind_, ptr);
+  #ifdef MEMKIND_IN_USE
+    memkind_free((struct memkind*)pmem_kind_, ptr);
+  #else
+    free(ptr);
+  #endif
   used_ -= size;
 }
