+++ libjiffy/src/jiffy/storage/block_memory_manager.h
@@ -5,7 +5,6 @@
 #include <cstddef>
 #include <atomic>
 #include <new>
-#include <memkind.h>
 #include <string>
 
 namespace jiffy {
@@ -26,7 +25,7 @@ class block_memory_manager {
    */
   explicit block_memory_manager(size_t capacity = 134217728,
                                 const std::string memory_mode = "DRAM",
-                                struct memkind *pmem_kind = nullptr);
+                                void* pmem_kind = nullptr);
 
   /**
    * @brief Allocate memory.
@@ -82,7 +81,7 @@ class block_memory_manager {
   size_t capacity_;
   std::atomic<size_t> used_;
   std::string memory_mode_;
-  struct memkind *pmem_kind_;
+  void* pmem_kind_;
 };
 
 }
