+++ libjiffy/test/chain_replication_test.cpp
@@ -33,7 +33,7 @@ TEST_CASE("chain_replication_no_failure_test", "[put][get]") {
                                                   STORAGE_MANAGEMENT_PORT_N(i));
     alloc->add_blocks(block_names[i]);
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = test_utils::init_pmem_kind();
+    void* pmem_kind = test_utils::init_kind();
     blocks[i] = test_utils::init_hash_table_blocks(block_names[i], memory_mode, pmem_kind);
 
     management_servers[i] = storage_management_server::create(blocks[i], HOST, STORAGE_MANAGEMENT_PORT_N(i));
@@ -119,7 +119,7 @@ TEST_CASE("chain_replication_head_failure_test", "[put][get]") {
                                                   STORAGE_MANAGEMENT_PORT_N(i));
     alloc->add_blocks(block_names[i]);
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = test_utils::init_pmem_kind();
+    void* pmem_kind = test_utils::init_kind();
     blocks[i] = test_utils::init_hash_table_blocks(block_names[i], memory_mode, pmem_kind);
 
     management_servers[i] = storage_management_server::create(blocks[i], HOST, STORAGE_MANAGEMENT_PORT_N(i));
@@ -215,7 +215,7 @@ TEST_CASE("chain_replication_mid_failure_test", "[put][get]") {
                                                   STORAGE_MANAGEMENT_PORT_N(i));
     alloc->add_blocks(block_names[i]);
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = test_utils::init_pmem_kind();
+    void* pmem_kind = test_utils::init_kind();
     blocks[i] = test_utils::init_hash_table_blocks(block_names[i], memory_mode, pmem_kind);
 
     management_servers[i] = storage_management_server::create(blocks[i], HOST, STORAGE_MANAGEMENT_PORT_N(i));
@@ -311,7 +311,7 @@ TEST_CASE("chain_replication_tail_failure_test", "[put][get]") {
                                                   STORAGE_MANAGEMENT_PORT_N(i));
     alloc->add_blocks(block_names[i]);
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = test_utils::init_pmem_kind();
+    void* pmem_kind = test_utils::init_kind();
     blocks[i] = test_utils::init_hash_table_blocks(block_names[i], memory_mode, pmem_kind);
 
     management_servers[i] = storage_management_server::create(blocks[i], HOST, STORAGE_MANAGEMENT_PORT_N(i));
@@ -407,7 +407,7 @@ TEST_CASE("chain_replication_add_block_test", "[put][get]") {
                                                   STORAGE_MANAGEMENT_PORT_N(i));
     alloc->add_blocks(block_names[i]);
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = test_utils::init_pmem_kind();
+    void* pmem_kind = test_utils::init_kind();
     blocks[i] = test_utils::init_hash_table_blocks(block_names[i], memory_mode, pmem_kind);
 
     management_servers[i] = storage_management_server::create(blocks[i], HOST, STORAGE_MANAGEMENT_PORT_N(i));
