+++ libjiffy/test/fifo_queue_local_partition_test.cpp
@@ -6,7 +6,6 @@
 #include <string>
 #include <iostream>
 #include <stdio.h>
-#include <memkind.h>
 #include "test_utils.h"
 
 using namespace ::jiffy::storage;
@@ -15,7 +14,7 @@ using namespace ::jiffy::persistent;
 
 TEST_CASE("fifo_queue_local_enqueue_read_dequeue_csv_test", "[enqueue][dequeue]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -56,7 +55,7 @@ TEST_CASE("fifo_queue_local_enqueue_read_dequeue_csv_test", "[enqueue][dequeue]"
 
 TEST_CASE("fifo_queue_local_enqueue_read_dequeue_binary_test", "[enqueue][dequeue]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
