+++ libjiffy/test/file_client_test.cpp
@@ -31,7 +31,7 @@ TEST_CASE("file_client_write_read_seek_test", "[write][read][seek]") {
   auto block_names = test_utils::init_block_names(NUM_BLOCKS, STORAGE_SERVICE_PORT, STORAGE_MANAGEMENT_PORT);
   alloc->add_blocks(block_names);
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   auto blocks = test_utils::init_file_blocks(block_names, memory_mode, pmem_kind, 134217728);
 
   auto storage_server = block_server::create(blocks, STORAGE_SERVICE_PORT);
@@ -95,7 +95,7 @@ TEST_CASE("file_client_concurrent_write_read_seek_test", "[write][read][seek]") 
   auto block_names = test_utils::init_block_names(20, STORAGE_SERVICE_PORT, STORAGE_MANAGEMENT_PORT);
   alloc->add_blocks(block_names);
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   auto blocks = test_utils::init_file_blocks(block_names, memory_mode, pmem_kind, BLOCK_SIZE);
   auto storage_server = block_server::create(blocks, STORAGE_SERVICE_PORT);
   std::thread storage_serve_thread([&storage_server] { storage_server->serve(); });
