+++ libjiffy/test/file_local_partition_test.cpp
@@ -4,7 +4,6 @@
 #include <vector>
 #include <string>
 #include <cstdio>
-#include <memkind.h>
 #include "test_utils.h"
 
 using namespace ::jiffy::storage;
@@ -12,7 +11,7 @@ using namespace ::jiffy::persistent;
 
 TEST_CASE("file_ls_write_read_csv_test", "[write][read]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -60,7 +59,7 @@ TEST_CASE("file_ls_write_read_csv_test", "[write][read]") {
 
 TEST_CASE("file_ls_write_read_binary_test", "[write][read]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
