+++ libjiffy/test/hash_table_local_partition_test.cpp
@@ -2,7 +2,6 @@
 #include "jiffy/storage/hashtable/hash_slot.h"
 #include "jiffy/storage/hashtable/hash_table_ops.h"
 #include "jiffy/storage/hashtable/hash_table_partition.h"
-#include <memkind.h>
 #include "test_utils.h"
 
 using namespace ::jiffy::storage;
@@ -10,7 +9,7 @@ using namespace ::jiffy::persistent;
 
 TEST_CASE("hash_table_ls_put_get_csv_test", "[put][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -49,7 +48,7 @@ TEST_CASE("hash_table_ls_put_get_csv_test", "[put][get]") {
 
 TEST_CASE("hash_table_ls_put_update_get_csv_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -91,7 +90,7 @@ TEST_CASE("hash_table_ls_put_update_get_csv_test", "[put][update][get]") {
 
 TEST_CASE("hash_table_ls_put_upsert_get_csv_test", "[put][upsert][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -133,7 +132,7 @@ TEST_CASE("hash_table_ls_put_upsert_get_csv_test", "[put][upsert][get]") {
 
 TEST_CASE("hash_table_ls_put_exists_remove_exists_csv_test", "[put][exists][remove][exists]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -175,7 +174,7 @@ TEST_CASE("hash_table_ls_put_exists_remove_exists_csv_test", "[put][exists][remo
 
 TEST_CASE("hash_table_ls_put_remove_get_csv_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -218,7 +217,7 @@ TEST_CASE("hash_table_ls_put_remove_get_csv_test", "[put][update][get]") {
 
 TEST_CASE("hash_table_ls_put_get_binary_test", "[put][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -256,7 +255,7 @@ TEST_CASE("hash_table_ls_put_get_binary_test", "[put][get]") {
 
 TEST_CASE("hash_table_ls_put_update_get_binary_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -298,7 +297,7 @@ TEST_CASE("hash_table_ls_put_update_get_binary_test", "[put][update][get]") {
 
 TEST_CASE("hash_table_ls_put_upsert_get_binary_test", "[put][upsert][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -340,7 +339,7 @@ TEST_CASE("hash_table_ls_put_upsert_get_binary_test", "[put][upsert][get]") {
 
 TEST_CASE("hash_table_ls_put_exists_remove_exists_binary_test", "[put][exists][remove][exists]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
@@ -382,7 +381,7 @@ TEST_CASE("hash_table_ls_put_exists_remove_exists_binary_test", "[put][exists][r
 
 TEST_CASE("hash_table_ls_put_remove_get_binary_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   property_map conf;
