+++ libjiffy/test/hash_table_partition_test.cpp
@@ -9,7 +9,7 @@ using namespace ::jiffy::persistent;
 
 TEST_CASE("hash_table_put_get_test", "[put][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
@@ -34,7 +34,7 @@ TEST_CASE("hash_table_put_get_test", "[put][get]") {
 
 TEST_CASE("hash_table_put_update_get_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
@@ -70,7 +70,7 @@ TEST_CASE("hash_table_put_update_get_test", "[put][update][get]") {
 
 TEST_CASE("hash_table_put_upsert_get_test", "[put][upsert][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
@@ -100,7 +100,7 @@ TEST_CASE("hash_table_put_upsert_get_test", "[put][upsert][get]") {
 
 TEST_CASE("hash_table_put_exists_remove_exists_test", "[put][exists][remove][exists]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
@@ -129,7 +129,7 @@ TEST_CASE("hash_table_put_exists_remove_exists_test", "[put][exists][remove][exi
 
 TEST_CASE("hash_table_put_remove_get_test", "[put][update][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
@@ -160,7 +160,7 @@ TEST_CASE("hash_table_put_remove_get_test", "[put][update][get]") {
 
 TEST_CASE("hash_table_storage_size_test", "[put][size][storage_size][reset]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
@@ -176,7 +176,7 @@ TEST_CASE("hash_table_storage_size_test", "[put][size][storage_size][reset]") {
 
 TEST_CASE("hash_table_flush_load_test", "[put][sync][reset][load][get]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   hash_table_partition block(&manager);
