+++ libjiffy/test/local_store_test.cpp
@@ -15,7 +15,7 @@ binary make_binary(const std::string& str, const block_memory_allocator<uint8_t>
 
 TEST_CASE("local_write_test", "[write]") {
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   block_memory_allocator<uint8_t> binary_allocator(&manager);
@@ -38,7 +38,7 @@ TEST_CASE("local_read_test", "[read]") {
   out << "key,value\n";
   out.close();
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   size_t capacity = 134217728;
   block_memory_manager manager(capacity, memory_mode, pmem_kind);
   block_memory_allocator<kv_pair_type> allocator(&manager);
