+++ libjiffy/test/notification_test.cpp
@@ -35,7 +35,7 @@ TEST_CASE("notification_test", "[subscribe][get_message]") {
                                                   STORAGE_MANAGEMENT_PORT);
   alloc->add_blocks(block_names);
   std::string memory_mode = getenv("JIFFY_TEST_MODE");
-  struct memkind* pmem_kind = test_utils::init_pmem_kind();
+  void* pmem_kind = test_utils::init_kind();
   auto blocks = test_utils::init_hash_table_blocks(block_names, memory_mode, pmem_kind);
 
   auto storage_server = block_server::create(blocks, STORAGE_SERVICE_PORT);
