+++ libjiffy/test/test_utils.h
@@ -5,7 +5,6 @@
 #include <vector>
 #include <iostream>
 #include <limits.h>
-#include <memkind.h>
 #include <thrift/transport/TBufferTransports.h>
 #include <thrift/transport/TSocket.h>
 #include "jiffy/storage/storage_management_ops.h"
@@ -14,6 +13,7 @@
 #include "jiffy/storage/hashtable/hash_table_partition.h"
 #include "jiffy/storage/notification/subscription_map.h"
 #include "jiffy/utils/logger.h"
+#include "jiffy/utils/mem_utils.h"
 #include "jiffy/directory/directory_ops.h"
 
 class dummy_storage_manager : public jiffy::storage::storage_management_ops {
@@ -239,7 +239,7 @@ class test_utils {
 
   static std::vector<std::shared_ptr<jiffy::storage::block>> init_hash_table_blocks(const std::vector<std::string> &block_ids,
                                                                                     std::string memory_mode = "DRAM",
-                                                                                    struct memkind* pmem_kind = nullptr,
+                                                                                    void* pmem_kind = nullptr,
                                                                                     size_t block_capacity = 134217728,
                                                                                     double threshold_lo = 0.05,
                                                                                     double threshold_hi = 0.95,                                          
@@ -277,7 +277,7 @@ class test_utils {
 
   static std::vector<std::shared_ptr<jiffy::storage::block>> init_file_blocks(const std::vector<std::string> &block_ids,
                                                                               std::string memory_mode = "DRAM",
-                                                                              struct memkind* pmem_kind = nullptr,
+                                                                              void* pmem_kind = nullptr,
                                                                               size_t block_capacity = 134217728,
                                                                               double threshold_lo = 0.05,
                                                                               double threshold_hi = 0.95,                                    
@@ -315,7 +315,7 @@ class test_utils {
 
   static std::vector<std::shared_ptr<jiffy::storage::block>> init_shared_log_blocks(const std::vector<std::string> &block_ids,
                                                                               std::string memory_mode = "DRAM",
-                                                                              struct memkind* pmem_kind = nullptr,
+                                                                              void* pmem_kind = nullptr,
                                                                               size_t block_capacity = 134217728,
                                                                               double threshold_lo = 0.05,
                                                                               double threshold_hi = 0.95,
@@ -352,7 +352,7 @@ class test_utils {
 
   static std::vector<std::shared_ptr<jiffy::storage::block>> init_fifo_queue_blocks(const std::vector<std::string> &block_ids,
                                                                                     std::string memory_mode = "DRAM",
-                                                                                    struct memkind* pmem_kind = nullptr,
+                                                                                    void* pmem_kind = nullptr,
                                                                                     size_t block_capacity = 134217728,
                                                                                     double threshold_lo = 0.05,
                                                                                     double threshold_hi = 0.95,                                          
@@ -372,28 +372,11 @@ class test_utils {
     return blks;
   }
 
-  static struct memkind* init_pmem_kind() {
+  static void* init_kind() {
     std::string pmem_path = getenv("PMEM_PATH"); 
     std::string memory_mode = getenv("JIFFY_TEST_MODE");
-    struct memkind* pmem_kind = nullptr;
-    if (memory_mode == "PMEM") {
-      size_t err = memkind_create_pmem(pmem_path.c_str(),0,&pmem_kind);
-      if(err) {
-        char error_message[MEMKIND_ERROR_MESSAGE_SIZE];
-        memkind_error_message(err, error_message, MEMKIND_ERROR_MESSAGE_SIZE);
-        fprintf(stderr, "%s\n", error_message);
-      }
-    }
+    void* pmem_kind = mem_utils::init_kind(memory_mode, pmem_path);
     return pmem_kind;
-  }
-
-  static void destroy_kind(struct memkind* pmem_kind) {
-    int err = memkind_destroy_kind(pmem_kind);
-    if(err) {
-      char error_message[MEMKIND_ERROR_MESSAGE_SIZE];
-      memkind_error_message(err, error_message, MEMKIND_ERROR_MESSAGE_SIZE);
-      fprintf(stderr, "%s\n", error_message);
-    }
   }
 
 };
