+++ storage/CMakeLists.txt
@@ -9,9 +9,13 @@ add_executable(storaged src/storage_server.cpp
         src/server_storage_tracker.cpp
         src/server_storage_tracker.h)
 
-add_dependencies(storaged boost_ep memkind_ep thrift_ep)
-
-target_link_libraries(storaged jiffy ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
+if (BUILD_MEMKIND_SUPPORT)
+  add_dependencies(storaged boost_ep memkind_ep thrift_ep)
+  target_link_libraries(storaged jiffy ${MEMKIND_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
+else()
+  add_dependencies(storaged boost_ep jemalloc_ep thrift_ep)
+  target_link_libraries(storaged jiffy ${JEMALLOC_LIBRARY} ${Boost_PROGRAM_OPTIONS_LIBRARY})
+endif()
 
 install(TARGETS storaged
         RUNTIME DESTINATION bin
